Add trigger ability for batcher
Triggering is based on latest peer activity in both
directions. This means that any packet received from
or sent to the peer will increase [rx|tx]_bytes
variable in WireGuard and in turn will trigger the timestamp
to change(if the value has changed).

This detection of changed is then used to trigger the batcher.

Batcher, instead of sleeping util the nearest
deadline and then batching will consult the trigger timestamp
to see if anything happened recently. If so it will batch immediately.

Warning: it should be pretty safe while there are good threshold added.
Given too big thresholds(bigger than intervals) will make peers trigger
each other and spiral out of control.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>